refactor: move modal to shared components
